move BlazegraphSlowQuery
