+++ delta/plugins/blazegraph/src/main/scala/ch/epfl/bluebrain/nexus/delta/plugins/blazegraph/slowqueries/model/BlazegraphSlowQuery.scala
@@ -1,4 +1,4 @@
-package ch.epfl.bluebrain.nexus.delta.plugins.blazegraph.slowqueries
+package ch.epfl.bluebrain.nexus.delta.plugins.blazegraph.slowqueries.model
 
 import ch.epfl.bluebrain.nexus.delta.rdf.IriOrBNode.Iri
 import ch.epfl.bluebrain.nexus.delta.rdf.query.SparqlQuery
@@ -7,13 +7,13 @@ import ch.epfl.bluebrain.nexus.delta.sourcing.model.Identity.Subject
 import ch.epfl.bluebrain.nexus.delta.sourcing.model.ProjectRef
 import doobie.Read
 import io.circe.Json
-
-import java.time.Instant
-import scala.concurrent.duration.{DurationLong, FiniteDuration}
 import doobie.postgres.implicits._
 import ch.epfl.bluebrain.nexus.delta.sourcing.implicits._
 
-case class BlazegraphSlowQuery(
+import java.time.Instant
+import scala.concurrent.duration.{DurationLong, FiniteDuration}
+
+final case class BlazegraphSlowQuery(
     view: ViewRef,
     query: SparqlQuery,
     duration: FiniteDuration,
